Final APP with Authentication & Authorization

+++ controllers/contactController.js
@@ -3,16 +3,16 @@ const Contact = require('../models/contactModel');
 
 //  @desc   Get all contacts
 //  @route  GET   /api/contacts
-//  @access public
+//  @access private
 
 const getContacts = asyncHandler(async (req, res) => {
-  const contacts = await Contact.find();
+  const contacts = await Contact.find({ user_id: req.user.id });
   res.status(200).json({ contacts });
 });
 
 //  @desc   Create new contact
 //  @route  POST   /api/contacts
-//  @access public
+//  @access private
 
 const createContact = asyncHandler(async (req, res) => {
   const { name, email, phone } = req.body;
@@ -20,14 +20,14 @@ const createContact = asyncHandler(async (req, res) => {
     res.status(400);
     throw new Error('All fields are required!');
   }
-  const contact = await Contact.create({ name, email, phone });
+  const contact = await Contact.create({ name, email, phone, user_id: req.user.id });
 
   res.status(201).json({ contact });
 });
 
 //  @desc   Get contact
 //  @route  GET   /api/contacts/:id
-//  @access public
+//  @access private
 
 const getContact = asyncHandler(async (req, res) => {
   const contact = await Contact.findById(req.params.id);
@@ -40,7 +40,7 @@ const getContact = asyncHandler(async (req, res) => {
 
 //  @desc   Update contact
 //  @route  PUT  /api/contacts/:id
-//  @access public
+//  @access private
 
 const updateContact = asyncHandler(async (req, res) => {
   const contact = await Contact.findById(req.params.id);
@@ -48,6 +48,11 @@ const updateContact = asyncHandler(async (req, res) => {
     res.status(404);
     throw new Error('Contact not found');
   }
+  if (contact.user_id.toString() !== req.user.id) {
+    res.status(403);
+    throw new Error('User have not permission to update other user contacts');
+  }
+
   const updatedContact = await Contact.findByIdAndUpdate(req.params.id, req.body, { new: true });
 
   res.status(200).json({ updatedContact });
@@ -55,7 +60,7 @@ const updateContact = asyncHandler(async (req, res) => {
 
 //  @desc   Delete contact
 //  @route  DELETE  /api/contacts/:id
-//  @access public
+//  @access private
 
 const deleteContact = asyncHandler(async (req, res) => {
   const contact = await Contact.findById(req.params.id);
@@ -63,6 +68,12 @@ const deleteContact = asyncHandler(async (req, res) => {
     res.status(404);
     throw new Error('Contact not found');
   }
+
+  if (contact.user_id.toString() !== req.user.id) {
+    res.status(403);
+    throw new Error('User have not permission to delete other user contacts');
+  }
+
   await Contact.findByIdAndDelete(req.params.id);
 
   res.status(200).json({ contact });
